rename server publication pdf server handler for clarity
