+++ publication-pdf-builder/publication-pdf-handler.js
